+++ src/components/Sign/Signup/signup.styled.js
@@ -1,9 +1,0 @@
-import { styled } from '@mui/styles';
-import { Box } from '@mui/material';
-
-export const StyledBox = styled(Box)({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  padding: '10px',
-});