+++ src/components/auth/Login/KakaoLogin.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import { setUser } from '../../../store/features/userSlice';
 import { Box } from '@mui/system';
 
 export default function KaKaoLogin() {
@@ -14,6 +13,7 @@ export default function KaKaoLogin() {
     Kakao.Auth.createLoginButton({
       container: '#kakao-login',
       success: function (response: string) {
+        console.log(response);
         setIsLogin(true);
         // 리다이렉션 처리 논의 필요
         navigate(-1);
@@ -30,7 +30,6 @@ export default function KaKaoLogin() {
         url: '/v2/user/me',
         success: function (response: string) {
           // 데이터 어떻게 저장할지 논의 필요
-          dispatch(setUser(response));
         },
         fail: function (error: string) {
           console.log(error);
@@ -41,5 +40,5 @@ export default function KaKaoLogin() {
 
   if (!Kakao) return null;
 
-  return <Box id="kakao-login" sx={{ m: 2, px: 6 }} />;
+  return <Box id="kakao-login" sx={{ width: '80%%', height: '80%', p: 1, mt: '0.5px' }} />;
 }