+++ src/components/auth/SignUp/signup.styled.ts
@@ -1,0 +1,8 @@
+import { styled } from '@mui/styles';
+
+export const StyledRow = styled('div')({
+  display: 'flex',
+  alignItems: 'center',
+  width: '100%',
+  margin: '8px 0 8px 0',
+});