+++ src/constants/EnvContant.ts
@@ -1,4 +1,7 @@
-export const MAP_API_KEY = process.env['REACT_APP_MAP_API_KEY'] as string;
 export const MOCK_API_URL = 'http://localhost:4000/api/v1' as string;
-export const KAKAO_API_KEY = process.env['REACT_APP_KAKAO_API_KEY'] as string;
+export const API_URL = 'https://react-node-project-butler-api.herokuapp.com/v1/web';
+
+export const MAP_API_KEY = process.env['REACT_APP_MAP_API_KEY'] || ('' as string);
+export const KAKAO_API_KEY = process.env['REACT_APP_KAKAO_API_KEY'] || ('' as string);
+export const GOOGLE_API_KEY = process.env['REACT_APP_GOOGLE_API_KEY'] || ('' as string);
 export const PAYPAL_CLIENT_ID = process.env['REACT_PAYPAL_CLIENT_ID'] as string;