+++ src/index.jsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
 import ReactDom from 'react-dom';
 import App from './App';
-import { store } from './store';
+import { persistor, store } from './store';
 import { Provider } from 'react-redux';
 import { ThemeProvider } from '@mui/material/styles';
 import { theme } from './styles/theme';
+import { PersistGate } from 'redux-persist/integration/react';
 import CssBaseline from '@mui/material/CssBaseline';
 
 // TODO:: 목서비스 워커 등록
@@ -19,10 +20,12 @@ if (process.env.NODE_ENV === 'development') {
 
 ReactDom.render(
   <Provider store={store}>
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <App />
-    </ThemeProvider>
+    <PersistGate loading={null} persistor={persistor}>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <App />
+      </ThemeProvider>
+    </PersistGate>
   </Provider>,
   document.querySelector('#root'),
 );