+++ src/store/service/auth.api.ts
@@ -1,13 +1,15 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-// import { API_URL } from '../../constants/EnvContant';
+import { MOCK_API_URL } from '../../constants/EnvContant';
+import { LoginRequest, LoginResponse } from '../../type/auth.type';
+
 
 export const authAPI = createApi({
   reducerPath: 'authAPI',
   baseQuery: fetchBaseQuery({
-    // baseUrl: `${API_URL}/auth`,
+    baseUrl: `${MOCK_API_URL}/auth`,
   }),
   endpoints: (build) => ({
-    login: build.mutation({
+    login: build.mutation<LoginResponse, LoginRequest>({
       query: ({ email, password }: { password: string; email: string }) => ({
         url: '/login',
         headers: { 'cors-free-key': '8ElvFN0JnirIOD4F' },